+++ lib/fulfil_api.rb
@@ -12,6 +12,7 @@ require "active_support/core_ext/hash/deep_merge"
 require "active_support/core_ext/hash/indifferent_access"
 require "active_support/core_ext/module/delegation"
 require "active_support/core_ext/object/blank"
+require "active_support/notifications"
 
 module FulfilApi
 end